clean up syntax
